@@ -1,50 +1,81 @@
 {
-    "name": "rpkamp/mailhog-behat-extension",
-    "description": "Mailhog Extension for Behat",
+    "name": "libresign/mailpit-behat-extension",
+    "description": "Mailpit Extension for Behat - Modern replacement for Mailhog",
     "type": "library",
     "license": "MIT",
     "authors": [
         {
+            "name": "LibreSign Team",
+            "email": "hello@libresign.coop"
+        },
+        {
             "name": "Remon van de Kamp",
-            "email": "rpkamp@gmail.com"
+            "email": "rpkamp@gmail.com",
+            "role": "Original Mailhog Behat extension author"
         }
     ],
     "autoload": {
         "psr-4": {
-            "rpkamp\\Behat\\MailhogExtension\\": "src/"
-        }
+            "LibreSign\\Behat\\MailpitExtension\\": "src/"
+        },
+        "files": [
+            "config/bootstrap.php"
+        ]
     },
     "autoload-dev": {
         "psr-4": {
-            "rpkamp\\Behat\\MailhogExtension\\Tests\\": "tests/"
+            "LibreSign\\Behat\\MailpitExtension\\Tests\\": "tests/"
         }
     },
     "minimum-stability": "stable",
     "require": {
-        "php": "^8.1",
+        "php": "^8.2",
         "behat/behat": "^3.8.0",
-        "rpkamp/mailhog-client": "^2.0",
-        "symfony/dependency-injection": "^6.4 || ^7.0",
+        "libresign/mailpit-client": "^0.1",
+        "nyholm/psr7": "^1.3",
         "psr-discovery/http-client-implementations": "^1.0",
         "psr-discovery/http-factory-implementations": "^1.0",
-        "symfony/http-client": "^6.0"
+        "symfony/dependency-injection": "^7.0",
+        "symfony/dotenv": "^7.0",
+        "symfony/http-client": "^7.0"
     },
     "require-dev": {
-        "guzzlehttp/psr7": "^1.6",
-        "php-parallel-lint/php-parallel-lint": "^1.2",
-        "swiftmailer/swiftmailer": "^6.2",
-        "phpunit/phpunit": "^9.0",
-        "mockery/mockery": "^1.3.3",
-        "phpstan/phpstan": "^1.1",
-        "phpstan/phpstan-mockery": "^1.1",
-        "doctrine/coding-standard": "^8.0",
-        "egulias/email-validator": "^2.1.23",
-        "phpmd/phpmd": "^2.9.1",
-        "nyholm/psr7": "^1.3"
+        "bamarni/composer-bin-plugin": "^1.8",
+        "roave/security-advisories": "dev-master",
+        "symfony/mailer": "^7.0",
+        "symfony/mime": "^7.0"
     },
     "config": {
+        "optimize-autoloader": true,
+        "classmap-authoritative": true,
+        "sort-packages": true,
         "allow-plugins": {
-            "dealerdirect/phpcodesniffer-composer-installer": true
+            "php-http/discovery": true,
+            "dealerdirect/phpcodesniffer-composer-installer": true,
+            "bamarni/composer-bin-plugin": true
         }
+    },
+    "extra": {
+        "bamarni-bin": {
+            "bin-links": true,
+            "forward-command": true
+        }
+    },
+    "scripts": {
+        "lint": "find . -name \\*.php -not -path './vendor/*' -not -path './vendor-bin/*' -print0 | xargs -0 -n1 php -l",
+        "cs:check": "vendor-bin/phpcs/vendor/squizlabs/php_codesniffer/bin/phpcs --standard=PSR2 src/",
+        "cs:fix": "vendor-bin/phpcs/vendor/squizlabs/php_codesniffer/bin/phpcbf --standard=PSR2 src/",
+        "phpmd": "vendor-bin/phpmd/vendor/phpmd/phpmd/src/bin/phpmd src text phpmd.xml",
+        "phpstan": "vendor-bin/phpstan/vendor/phpstan/phpstan/phpstan analyze src/",
+        "psalm": "vendor-bin/psalm/vendor/vimeo/psalm/psalm --no-cache --threads=$(nproc)",
+        "psalm:update-baseline": "vendor-bin/psalm/vendor/vimeo/psalm/psalm --threads=$(nproc) --update-baseline --set-baseline=tests/psalm-baseline.xml",
+        "test:unit": "vendor-bin/phpunit/vendor/phpunit/phpunit/phpunit -c phpunit.xml.dist --testsuite unit --colors=always",
+        "test:behat": "vendor/bin/behat --no-interaction --profile=default --format=progress",
+        "post-install-cmd": [
+          "@composer bin all install --ansi"
+        ],
+        "post-update-cmd": [
+          "@composer bin all update --ansi"
+        ]
     }
 }
