@@ -1,24 +1,25 @@
 <?php
 declare(strict_types=1);
 
-namespace rpkamp\Behat\MailhogExtension\Tests\ServiceContainer;
+namespace LibreSign\Behat\MailpitExtension\Tests\ServiceContainer;
 
 use Behat\Behat\Context\ServiceContainer\ContextExtension;
 use Behat\Testwork\EventDispatcher\ServiceContainer\EventDispatcherExtension;
+use LibreSign\Behat\MailpitExtension\Context\Initializer\MailpitAwareInitializer;
+use LibreSign\Behat\MailpitExtension\Listener\EmailPurgeListener;
+use LibreSign\Behat\MailpitExtension\ServiceContainer\MailpitExtension;
+use LibreSign\Mailpit\MailpitClient;
+use PHPUnit\Framework\Attributes\Test;
 use PHPUnit\Framework\TestCase;
 use Psr\Http\Message\RequestFactoryInterface;
 use Psr\Http\Message\StreamFactoryInterface;
-use rpkamp\Behat\MailhogExtension\Context\Initializer\MailhogAwareInitializer;
-use rpkamp\Behat\MailhogExtension\Listener\EmailPurgeListener;
-use rpkamp\Behat\MailhogExtension\ServiceContainer\MailhogExtension;
-use rpkamp\Mailhog\MailhogClient;
 use Symfony\Component\Config\Definition\Builder\ArrayNodeDefinition;
 use Symfony\Component\Config\Definition\Exception\InvalidConfigurationException;
 use Symfony\Component\Config\Definition\Processor;
 use Symfony\Component\DependencyInjection\ContainerBuilder;
 use Symfony\Component\HttpClient\Psr18Client;
 
-final class MailhogExtensionTest extends TestCase
+final class MailpitExtensionTest extends TestCase
 {
     public const BASE_URL = 'http://localhost:10025/';
 
@@ -29,101 +30,83 @@ final class MailhogExtensionTest extends TestCase
         $this->container = new ContainerBuilder();
     }
 
-    /**
-     * @test
-     */
+    #[Test]
     public function it_should_set_the_base_url_as_a_container_parameter(): void
     {
         $this->loadExtension($this->container);
 
-        $this->assertEquals(self::BASE_URL, $this->container->getParameter('mailhog.base_url'));
+        $this->assertEquals(self::BASE_URL, $this->container->getParameter('mailpit.base_url'));
     }
 
-    /**
-     * @test
-     */
+    #[Test]
     public function it_should_set_a_http_client_instance_in_the_container(): void
     {
         $this->loadExtension($this->container);
 
-        $this->assertContainerHasServiceOfClass(Psr18Client::class, 'mailhog.http_client');
+        $this->assertContainerHasServiceOfClass(Psr18Client::class, 'mailpit.http_client');
     }
 
-    /**
-     * @test
-     */
+    #[Test]
     public function it_should_set_a_http_stream_factory_instance_in_the_container(): void
     {
         $this->loadExtension($this->container);
 
-        $this->assertContainerHasServiceOfClass(StreamFactoryInterface::class, 'mailhog.http_stream_factory');
+        $this->assertContainerHasServiceOfClass(StreamFactoryInterface::class, 'mailpit.http_stream_factory');
     }
 
-    /**
-     * @test
-     */
+    #[Test]
     public function it_should_set_a_http_request_factory_in_the_container(): void
     {
         $this->loadExtension($this->container);
 
-        $this->assertContainerHasServiceOfClass(RequestFactoryInterface::class, 'mailhog.http_request_factory');
+        $this->assertContainerHasServiceOfClass(RequestFactoryInterface::class, 'mailpit.http_request_factory');
     }
 
-    /**
-     * @test
-     */
-    public function it_should_set_a_mailhog_client_instance_in_the_container(): void
+    #[Test]
+    public function it_should_set_a_mailpit_client_instance_in_the_container(): void
     {
         $this->loadExtension($this->container);
 
-        $this->assertContainerHasServiceOfClass(MailhogClient::class, 'mailhog.client');
+        $this->assertContainerHasServiceOfClass(MailpitClient::class, 'mailpit.client');
     }
 
-    /**
-     * @test
-     */
+    #[Test]
     public function it_should_set_initializer_with_correct_tag(): void
     {
         $this->loadExtension($this->container);
 
-        $this->assertContainerHasServiceOfClass(MailhogAwareInitializer::class, 'mailhog.context_initializer');
+        $this->assertContainerHasServiceOfClass(MailpitAwareInitializer::class, 'mailpit.context_initializer');
 
-        $definition = $this->container->getDefinition('mailhog.context_initializer');
+        $definition = $this->container->getDefinition('mailpit.context_initializer');
         $this->assertEquals([['priority' => 0]], $definition->getTag(ContextExtension::INITIALIZER_TAG));
     }
 
-    /**
-     * @test
-     */
+    #[Test]
     public function it_should_set_and_register_purge_listener(): void
     {
         $this->loadExtension($this->container);
 
-        $this->assertContainerHasServiceOfClass(EmailPurgeListener::class, 'mailhog.purge_listener');
+        $this->assertContainerHasServiceOfClass(EmailPurgeListener::class, 'mailpit.purge_listener');
 
-        $definition = $this->container->getDefinition('mailhog.purge_listener');
+        $definition = $this->container->getDefinition('mailpit.purge_listener');
         $this->assertEquals([['priority' => 0]], $definition->getTag(EventDispatcherExtension::SUBSCRIBER_TAG));
     }
 
-    /**
-     * @test
-     */
+    #[Test]
     public function it_should_throw_exception_when_no_base_url_supplied(): void
     {
         $node = new ArrayNodeDefinition(null);
-        (new MailhogExtension())->configure($node);
+        (new MailpitExtension())->configure($node);
 
         $this->expectException(InvalidConfigurationException::class);
         (new Processor())->process($node->getNode(), [[]]);
     }
 
-    /**
-     * @test
-     */
+    #[Test]
     public function it_should_set_default_email_purge_tag_if_none_supplied(): void
     {
         $node = new ArrayNodeDefinition(null);
-        (new MailhogExtension())->configure($node);
+        (new MailpitExtension())->configure($node);
 
         $configuration = (new Processor())->process($node->getNode(), [['base_url' => self::BASE_URL]]);
         $this->assertEquals('email', $configuration['purge_tag']);
@@ -137,7 +120,7 @@ final class MailhogExtensionTest extends TestCase
 
     private function loadExtension(ContainerBuilder $container): void
     {
-        $extension = new MailhogExtension();
+        $extension = new MailpitExtension();
         $extension->load($container, ['base_url' => self::BASE_URL, 'purge_tag' => 'email']);
     }
 }
