@@ -1,37 +1,38 @@
 <?php
 declare(strict_types=1);
 
-namespace rpkamp\Behat\MailhogExtension\Listener;
+namespace LibreSign\Behat\MailpitExtension\Listener;
 
 use Behat\Behat\EventDispatcher\Event\ExampleTested;
 use Behat\Behat\EventDispatcher\Event\ScenarioLikeTested;
 use Behat\Behat\EventDispatcher\Event\ScenarioTested;
-use rpkamp\Mailhog\MailhogClient;
+use LibreSign\Mailpit\MailpitClient;
 use Symfony\Component\EventDispatcher\EventSubscriberInterface;
 
 use function array_merge;
 
 final class EmailPurgeListener implements EventSubscriberInterface
 {
-    private MailhogClient $client;
+    private MailpitClient $client;
 
     /**
      * The tag name for scenarios/features that trigger
-     * a mailhog purge. Defaults to 'email'.
+     * a mailpit purge. Defaults to 'email'.
      *
      * @var string
      */
     private $purgeTag;
 
-    public function __construct(MailhogClient $client, string $purgeTag)
+    public function __construct(MailpitClient $client, string $purgeTag)
     {
         $this->client = $client;
         $this->purgeTag = $purgeTag;
     }
 
     /**
-     * @return array<string, array<string|int>>
+     * @return array<string, array{0?: array{0: string, 1?: int}|string, 1?: array{0: string, 1?: int}|int, ...<int<0, max>, array{0: string, 1?: int}>}|string>
      */
+    #[\Override]
     public static function getSubscribedEvents(): array
     {
         return [
